+++ tests/AnalyzeActionTrait.php
@@ -11,10 +11,9 @@ namespace DecodeLabs\Greenleaf\Tests;
 
 use DecodeLabs\Greenleaf\Action;
 use DecodeLabs\Greenleaf\ActionTrait;
+use DecodeLabs\Greenleaf\Request as LeafRequest;
 use DecodeLabs\Harvest\Response\Json as JsonResponse;
 use Psr\Http\Message\ResponseInterface as Response;
-use Psr\Http\Message\ServerRequestInterface as Request;
-use DecodeLabs\Singularity\Url\Leaf as LeafUrl;
 
 
 class AnalyzeActionTrait implements Action
@@ -22,9 +21,7 @@ class AnalyzeActionTrait implements Action
     use ActionTrait;
 
     public function execute(
-        Request $request,
-        LeafUrl $uri,
-        array $parameters
+        LeafRequest $request
     ): Response {
         return new JsonResponse(['foo' => 'bar']);
     }